pyproject.toml
[project]
dependencies = [
  "celery",
  "django",
  "mysqlclient",
  "pytz",
  "djangorestframework-jwt",
  "django-crispy-forms",
  "django-widget-tweaks",
  "crispy-bootstrap4",
  "celery[redis]",
  "Pillow",
  "python-dotenv",
]
description = "The latest version of HSEFZ digital services"
name = "hsefz-digital-life"
readme = "README.md"
requires-python = ">=3.12"
version = "3.0.0"
